requirements.txt
flask
flask[async]
requests
boto3
openai
python-telegram-bot
python-dotenv